Add Counter base class for models
