+++ SecretSanta/src/SecretSanta.Domain/Models/Group.cs
@@ -4,19 +4,13 @@ using SecretSanta.Domain.Interfaces;
 
 namespace SecretSanta.Domain.Models
 {
-	public class Group : IEntity
+	public class Group : Counter, IEntity
 	{
-		private static int IdCounter { get; set; }
-
-		public int Id { get; set; }
 		public string Title { get; set; }
 		public List<UserGroup> UserGroups { get; set; }
 
 		public Group(string title)
 		{
-			IdCounter++;
-
-			Id = IdCounter;
 			Title = title ?? throw new ArgumentNullException(nameof(title));
 			UserGroups = new List<UserGroup>();
 		}
@@ -24,11 +18,6 @@ namespace SecretSanta.Domain.Models
 		public override string ToString()
 		{
 			return $"{Id}: {Title}";
-		}
-
-		public static void ResetCounter()
-		{
-			IdCounter = 0;
 		}
 	}
 }
