+++ SecretSanta/src/SecretSanta.Domain/Models/Pairing.cs
@@ -3,20 +3,14 @@ using SecretSanta.Domain.Interfaces;
 
 namespace SecretSanta.Domain.Models
 {
-	public class Pairing : IEntity
+	public class Pairing : Counter, IEntity
 	{
-		private static int IdCounter { get; set; }
-
-		public int Id { get; set; }
 		public int RecipientId { get; set; }
 		public int SantaId { get; set; }
 		public int GroupId { get; set; }
 
 		public Pairing(int recipientId, int santaId, int groupId = 0)
 		{
-			IdCounter++;
-
-			Id = IdCounter;
 			RecipientId = recipientId > -1 ? recipientId : throw new ArgumentException(nameof(recipientId));
 			SantaId = santaId > -1 ? santaId : throw new ArgumentException(nameof(santaId));
 			GroupId = groupId > -1 ? groupId : throw new ArgumentException(nameof(groupId));
@@ -25,11 +19,6 @@ namespace SecretSanta.Domain.Models
 		public override string ToString()
 		{
 			return $"{Id}: Recipient: {RecipientId}, Santa: {SantaId}";
-		}
-
-		public static void ResetCounter()
-		{
-			IdCounter = 0;
 		}
 	}
 }
