+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/GiftServiceTests.cs
@@ -7,7 +7,7 @@ using SecretSanta.Domain.Services;
 namespace SecretSanta.Domain.Tests.Services
 {
 	[TestClass]
-	public class GiftServiceServiceTests : BaseServiceTest
+	public class GiftServiceTests : BaseServiceTest
 	{
 		[TestMethod]
 		[ExpectedException(typeof(ArgumentNullException))]
