+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/MessageServiceTests.cs
@@ -7,7 +7,7 @@ using SecretSanta.Domain.Services;
 namespace SecretSanta.Domain.Tests.Services
 {
 	[TestClass]
-	public class MessageServiceServiceTests : BaseServiceTest
+	public class MessageServiceTests : BaseServiceTest
 	{
 		[TestMethod]
 		[ExpectedException(typeof(ArgumentNullException))]
