+++ SecretSanta/test/SecretSanta.Domain.Tests/Services/UserGroupServiceTests.cs
@@ -9,7 +9,7 @@ using SecretSanta.Domain.Services;
 namespace SecretSanta.Domain.Tests.Services
 {
 	[TestClass]
-	public class UserGroupServiceServiceTests : BaseServiceTest
+	public class UserGroupServiceTests : BaseServiceTest
 	{
 		[TestMethod]
 		[ExpectedException(typeof(ArgumentNullException))]
